cc_fraud_detection/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
imbalanced-learn
pyyaml
torch
tqdm
